backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
sentence-transformers==3.3.1
google-generativeai==0.8.1
openai>=1.40.0
python-dotenv==1.0.1
gitpython==3.1.43
numpy>=2.1.0
